+++ database/music_album_schema.sql
