+++ managers/game_manager.rb
@@ -1,7 +1,7 @@
-require_relative 'game'
-require_relative 'author'
+require_relative '../classes/game'
+require_relative '../classes/author'
 require 'json'
-require_relative './data_storage_handler'
+require_relative '../modules/data_storage_handler'
 
 class GameManager
   attr_accessor :games
