+++ modules/data_storage_handler.rb
@@ -1,7 +1,7 @@
 require 'fileutils'
 require 'json'
-require_relative './music_album'
-require_relative './genre'
+require_relative '../classes/music_album'
+require_relative '../classes/genre'
 
 module DataStorageHandler
   def self.save_data(file_name, data)
